add mysql protocol infer

+++ crates/l7-parser/src/mysql/constants.rs
@@ -1,1 +1,1 @@
-pub(super) const MYSQL_HEADER_MIN_SIZE: u32 = 4;
+pub(super) const MYSQL_HEADER_MIN_SIZE: usize = 4;
